requirements.txt
Flask>=2.2
pymongo>=4.0
passlib[bcrypt]>=1.7
python-jose>=3.3
python-dotenv